+++ sdk/kubernetesconfiguration/azure-resourcemanager-kubernetesconfiguration/src/main/java/com/azure/resourcemanager/kubernetesconfiguration/models/MessageLevelType.java
@@ -8,7 +8,7 @@ import com.azure.core.util.ExpandableStringEnum;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import java.util.Collection;
 
-/** Defines values for MessageLevelType. */
+/** Level of the message. */
 public final class MessageLevelType extends ExpandableStringEnum<MessageLevelType> {
     /** Static value Error for MessageLevelType. */
     public static final MessageLevelType ERROR = fromString("Error");
@@ -18,6 +18,15 @@ public final class MessageLevelType extends ExpandableStringEnum<MessageLevelTyp
 
     /** Static value Information for MessageLevelType. */
     public static final MessageLevelType INFORMATION = fromString("Information");
+
+    /**
+     * Creates a new instance of MessageLevelType value.
+     *
+     * @deprecated Use the {@link #fromString(String)} factory method.
+     */
+    @Deprecated
+    public MessageLevelType() {
+    }
 
     /**
      * Creates or finds a MessageLevelType from its string representation.
@@ -30,7 +39,11 @@ public final class MessageLevelType extends ExpandableStringEnum<MessageLevelTyp
         return fromString(name, MessageLevelType.class);
     }
 
-    /** @return known MessageLevelType values. */
+    /**
+     * Gets known MessageLevelType values.
+     *
+     * @return known MessageLevelType values.
+     */
     public static Collection<MessageLevelType> values() {
         return values(MessageLevelType.class);
     }